+++ tests/Models/Author.php
@@ -6,9 +6,15 @@ use Illuminate\Database\Eloquent\Model;
  * Class Author
  *
  * A test model used for the relationship tests.
+ *
+ * @package Cviebrock\EloquentSluggable\Tests\Models
+ *
+ * @property integer id
+ * @property string name
  */
 class Author extends Model
 {
+
     /**
      * Indicates if the model should be timestamped.
      *