+++ tests/Models/PostWithMultipleSlugsAndCustomSlugKey.php
@@ -1,0 +1,11 @@
+<?php namespace Cviebrock\EloquentSluggable\Tests\Models;
+
+/**
+ * Class PostWithMultipleSlugsAndCustomSlugKey
+ */
+class PostWithMultipleSlugsAndCustomSlugKey extends PostWithMultipleSlugsAndHelperTrait
+{
+
+    protected $slugKeyName = 'dummy';
+    
+}