+++ tests/Models/PostWithMultipleSlugsAndHelperTrait.php
@@ -1,0 +1,13 @@
+<?php namespace Cviebrock\EloquentSluggable\Tests\Models;
+
+use Cviebrock\EloquentSluggable\SluggableScopeHelpers;
+
+/**
+ * Class PostWithMultipleSlugsAndPrimary
+ */
+class PostWithMultipleSlugsAndHelperTrait extends PostWithMultipleSlugs
+{
+
+    use SluggableScopeHelpers;
+
+}