+++ tests/Models/PostWithRelation.php
@@ -4,6 +4,8 @@
  * Class PostWithRelation
  *
  * A test model used for the relationship tests.
+ *
+ * @property \Cviebrock\EloquentSluggable\Tests\Models\Author author
  */
 class PostWithRelation extends Post
 {