+++ tests/PrimarySlugTests.php
@@ -1,10 +1,7 @@
 <?php namespace Cviebrock\EloquentSluggable\Tests;
 
-use Cviebrock\EloquentSluggable\Services\SlugService;
-use Cviebrock\EloquentSluggable\Tests\Models\Post;
-use Cviebrock\EloquentSluggable\Tests\Models\PostWithMultipleSlugsAndPrimary;
-use Cviebrock\EloquentSluggable\Tests\Models\PostWithMultipleSlugsAndPrimaryProperty;
-use Cviebrock\EloquentSluggable\Tests\Models\PostWithPrimarySlug;
+use Cviebrock\EloquentSluggable\Tests\Models\PostWithMultipleSlugsAndCustomSlugKey;
+use Cviebrock\EloquentSluggable\Tests\Models\PostWithMultipleSlugsAndHelperTrait;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 /**
@@ -16,14 +13,14 @@ class PrimarySlugTests extends TestCase
 {
 
     /**
-     * Test primary slug is set to first defined slug if $model->slugKeyName is not set
+     * Test that primary slug is set to $model->slugKeyName when set.
      *
      * @test
      */
     public function testSlugKeyNameProperty()
     {
 
-        $post = PostWithMultipleSlugsAndPrimaryProperty::create([
+        $post = PostWithMultipleSlugsAndCustomSlugKey::create([
             'title' => 'A Post Title',
             'subtitle' => 'A Post Subtitle'
         ]);
@@ -31,17 +28,16 @@ class PrimarySlugTests extends TestCase
         $this->assertEquals('dummy', $post->getSlugKeyName());
         $this->assertEquals('a.post.subtitle', $post->dummy);
         $this->assertEquals('a.post.subtitle', $post->getSlugKey());
-
     }
 
     /**
-     * Test primary slug is set to first defined slug if $model->slugKeyName is not set
+     * Test primary slug is set to first defined slug if $model->slugKeyName is not set.
      *
      * @test
      */
     public function testFirstSlugAsFallback()
     {
-        $post = PostWithMultipleSlugsAndPrimary::create([
+        $post = PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title'
         ]);
 
@@ -49,74 +45,88 @@ class PrimarySlugTests extends TestCase
         $this->assertEquals('a-post-title', $post->getSlugKey());
     }
 
-
     /**
-     * Test primary slug query scope
+     * Test primary slug query scope.
      *
      * @test
      */
     public function testQueryScope()
     {
 
-        PostWithMultipleSlugsAndPrimary::create([
+        PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title A'
         ]);
 
-        $post = PostWithMultipleSlugsAndPrimary::create([
+        $post = PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title B'
         ]);
 
-        PostWithMultipleSlugsAndPrimary::create([
+        PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title C'
         ]);
 
-        $this->assertEquals($post->id, PostWithMultipleSlugsAndPrimary::whereSlug('a-post-title-b')->first()->id);
-
+        $this->assertEquals($post->getKey(),
+            PostWithMultipleSlugsAndHelperTrait::whereSlug('a-post-title-b')->first()->getKey());
     }
 
+    /**
+     * Test finding a model by its primary slug.
+     *
+     * @test
+     */
     public function testFindBySlug()
     {
 
-        PostWithMultipleSlugsAndPrimary::create([
+        PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title A'
         ]);
 
-        $post = PostWithMultipleSlugsAndPrimary::create([
+        $post = PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title B'
         ]);
 
-        PostWithMultipleSlugsAndPrimary::create([
+        PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title C'
         ]);
 
-        $this->assertEquals($post->id, PostWithMultipleSlugsAndPrimary::findBySlug('a-post-title-b')->id);
+        $this->assertEquals($post->getKey(),
+            PostWithMultipleSlugsAndHelperTrait::findBySlug('a-post-title-b')->getKey());
     }
 
+    /**
+     * Test finding a model by its primary slug fails if the slug does not exist.
+     *
+     * @test
+     */
     public function testFindBySlugReturnsNullForNoRecord()
     {
-        $this->assertNull(PostWithMultipleSlugsAndPrimary::findBySlug('not a real record'));
+        $this->assertNull(PostWithMultipleSlugsAndHelperTrait::findBySlug('not a real record'));
     }
 
+    /**
+     * Test finding a model by its primary slug throws an exception if the slug does not exist.
+     *
+     * @test
+     */
     public function testFindBySlugOrFail()
     {
-        PostWithMultipleSlugsAndPrimary::create([
+        PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title A'
         ]);
 
-        $post = PostWithMultipleSlugsAndPrimary::create([
+        $post = PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title B'
         ]);
 
-        PostWithMultipleSlugsAndPrimary::create([
+        PostWithMultipleSlugsAndHelperTrait::create([
             'title' => 'A Post Title C'
         ]);
 
-        $this->assertEquals($post->id, PostWithMultipleSlugsAndPrimary::findBySlugOrFail('a-post-title-b')->id);
+        $this->assertEquals($post->getKey(),
+            PostWithMultipleSlugsAndHelperTrait::findBySlugOrFail('a-post-title-b')->getKey());
 
         $this->expectException(ModelNotFoundException::class);
 
-        PostWithMultipleSlugsAndPrimary::findBySlugOrFail('not a real record');
-
+        PostWithMultipleSlugsAndHelperTrait::findBySlugOrFail('not a real record');
     }
-
 }